Handler update + Reload fix
- Split up client and common sides of the ItemEventHandlers
- Fixed reload overlays and times for new RangedWeapon implementation
